requirements.txt
tensorflow
Flask
numpy
Pillow
